requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.5.0
decorator==5.1.1
edge-tts==7.0.0
exceptiongroup==1.2.2
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gTTS==2.5.4
huggingface-hub==0.29.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.20.1
openai-whisper==20240930
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.3
pysrt==1.1.2
pytest==8.3.4
python-dotenv==1.0.1
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sounddevice==0.5.1
srt==3.5.3
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
whisper-ctranslate2==0.5.2
yarl==1.18.3
bs4